backend/requirements.txt
flask
flask-cors
firebase-admin
google-generativeai
requests
python-dotenv
